requirements.txt
aiohttp==3.9.0b0
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.1.0
audioread==3.0.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
decorator==5.1.1
discord.py==2.3.2
einops==0.7.0
ffmpeg==1.4
filelock==3.9.0
frozenlist==1.4.0
fsspec==2023.10.0
grpcio==1.59.3
huggingface-hub==0.19.4
idna==3.4
inflect==7.0.0
Jinja2==3.1.2
joblib==1.3.2
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.41.1
MarkupSafe==2.1.3
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
networkx==3.0
numba==0.58.1
numpy==1.24.1
packaging==23.2
Pillow==9.3.0
platformdirs==4.0.0
pooch==1.8.0
progressbar==2.5
protobuf==4.25.1
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pydub==0.25.1
pyht==0.0.16
PyNaCl==1.5.0
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
rotary-embedding-torch==0.3.5
safetensors==0.4.1rc1
scikit-learn==1.3.2
scipy==1.11.4
soundfile==0.12.1
soxr==0.3.7
sympy==1.12
threadpoolctl==3.2.0
tokenizers==0.13.3
torch==2.1.1+cu118
torchaudio==2.1.1+cu118
torchvision==0.16.1+cu118
tortoise-tts==3.0.0
tqdm==4.66.1
transformers==4.31.0
typing_extensions==4.6.1
Unidecode==1.3.7
urllib3==2.1.0
yarl==1.9.3
